requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
jinja2>=3.1.4
python-multipart>=0.0.9
python-dotenv>=1.0.1
aiosqlite>=0.20.0
# Advanced DB integrations (for scaling):
# redis>=5.0.4
# supabase>=2.5.0
# pinecone-client>=4.1.1
google-generativeai>=0.5.4
requests>=2.32.2
beautifulsoup4>=4.12.3
python-whois>=0.8.0
